requirements.txt
asgiref==3.9.2
certifi==2025.8.3
cffi==1.17.1
crispy-bootstrap4==2025.6
cryptography==43.0.3
distlib==0.4.0
Django==5.2.7
django-crispy-forms==2.4
experta==1.9.4
filelock==3.19.1
frozendict==1.2
gunicorn==23.0.0
numpy==2.3.3
packaging==25.0
pillow==12.0.0
pipenv==2025.0.4
platformdirs==4.4.0
pycparser==2.22
schema==0.6.7
scikit-fuzzy==0.5.0
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
virtualenv==20.34.0
